feat: add json struct tags to PublicCodeV0
Add json struct tags so we can serialize the struct to JSON
and use it to pass the data around from the Go code to the various
language bindings as a simple char*, instead of wrapping the whole
type.

Unfortunately the yaml library doesn't allow us to use the same
tag, and we end up duplicating them:
https://github.com/go-yaml/yaml/issues/424

This duplication should be temporary, as goccy/go-yaml supports the
json tags for YAML serialization (https://github.com/italia/publiccode-parser-go/pull/221).